requirements.txt
openai
python-dotenv
pypdf
llama-index
streamlit